fix: client not aligned with server
