+++ packages/shared/src/hooks/useRefreshToken.ts
@@ -1,3 +1,5 @@
+'use client';
+
 import { useEffect } from 'react';
 import { differenceInMilliseconds } from 'date-fns';
 import type { AccessToken } from '../lib/boot';
